requirements.txt
asgiref==3.8.1
Django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pillow==11.0.0
PyJWT==2.10.1
rest-framework-generic-relations==2.2.0
sqlparse==0.5.3
typing_extensions==4.12.2
